% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawIOElement.R
\name{as.character.DrawIOElement}
\alias{as.character.DrawIOElement}
\title{Retrieve character representation of DrawIOElement}
\usage{
\method{as.character}{DrawIOElement}(x, include_declaration = FALSE, ...)
}
\arguments{
\item{x}{DrawIOElement object}

\item{include_declaration}{Include the XML declaration at the top? Default: FALSE}

\item{...}{other arguments}
}
\description{
Retrieve character representation of DrawIOElement
}
