% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawIOElement.R
\name{build_style}
\alias{build_style}
\title{Build a properly formatted style string}
\usage{
build_style(...)
}
\arguments{
\item{...}{named and unnamed arguments converted to a style string}
}
\description{
Drop NULLs
}
\examples{
{
\dontrun{
build_style(a = 1, b = 2)
}
}
}
