% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawIOElement.R
\name{coords_to_stencil}
\alias{coords_to_stencil}
\title{Convert a set of coordinates into a stencil for custom shape definition}
\usage{
coords_to_stencil(
  xs,
  ys,
  close = FALSE,
  stroke = c("stroke", "fill", "fillstroke")
)
}
\arguments{
\item{xs, ys}{coordinates}

\item{close}{is shape closed? default: FALSE (open)}

\item{stroke}{determines if stentil is filled, just the outline, or both}
}
\description{
Convert a set of coordinates into a stencil for custom shape definition
}
