% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.R
\name{decode}
\alias{decode}
\title{Decode a drawio encode string.}
\usage{
decode(encoded_string)
}
\arguments{
\item{encoded_string}{encode string in a draw io document}
}
\value{
raw text
}
\description{
Apply (1) base64 decoding, (2) gzip inflation, and (3) URL decode
}
\details{
For comparison see online tool: \url{https://jgraph.github.io/drawio-tools/tools/convert.html}
}
