% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.R
\name{encode}
\alias{encode}
\title{Encode a raw string into a encoded/compressed drawio string.}
\usage{
encode(raw_text)
}
\arguments{
\item{raw_text}{encode string in a draw io document}
}
\value{
encoded string
}
\description{
Apply (1) URL encode, (2) gzip deflation, (3) base64 encoding
}
\details{
For comparison see online tool: \url{https://jgraph.github.io/drawio-tools/tools/convert.html}
}
