% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colour.R
\name{r2hex}
\alias{r2hex}
\title{Convert any R colour to a hex colour}
\usage{
r2hex(x)
}
\arguments{
\item{x}{any R colour}
}
\value{
hex colour string
}
\description{
Return 'none' to indicate a transparent colour
}
